Release v0.13.0

+++ Api/Xipona.Api.Contracts/Items/Queries/Get/ItemAvailabilityContract.cs
@@ -1,8 +1,16 @@
 ﻿namespace ProjectHermes.Xipona.Api.Contracts.Items.Queries.Get
 {
+    /// <summary>
+    /// Represents availability of an item in a store.
+    /// </summary>
     public class ItemAvailabilityContract
     {
-        public ItemAvailabilityContract(ItemStoreContract store, float price,
+        /// <summary>
+        /// </summary>
+        /// <param name="store"></param>
+        /// <param name="price"></param>
+        /// <param name="defaultSection"></param>
+        public ItemAvailabilityContract(ItemStoreContract store, decimal price,
             ItemSectionContract defaultSection)
         {
             Store = store;
@@ -10,8 +18,19 @@
             DefaultSection = defaultSection;
         }
 
+        /// <summary>
+        /// The store where the item is available.
+        /// </summary>
         public ItemStoreContract Store { get; }
-        public float Price { get; }
+
+        /// <summary>
+        /// The item's price in the store.
+        /// </summary>
+        public decimal Price { get; }
+
+        /// <summary>
+        /// The section where the item is normally located in the store.
+        /// </summary>
         public ItemSectionContract DefaultSection { get; }
     }
 }
