+++ Frontend/Xipona.Frontend.Redux/ShoppingList/States/TemporaryItemCreator.cs
@@ -1,3 +1,3 @@
 ﻿namespace ProjectHermes.Xipona.Frontend.Redux.ShoppingList.States;
-public record TemporaryItemCreator(string ItemName, ShoppingListStoreSection? Section, float Price,
+public record TemporaryItemCreator(string ItemName, ShoppingListStoreSection? Section, decimal Price,
     int SelectedQuantityTypeId, bool IsButtonEnabled, bool IsOpen, bool IsSaving);
